Pull out Dialog components to their own files
